[ci] Fix validation of kiwi repositories
When the image configuration an kiwi repository belongs to has
changed, the record shouldn't be saved. The test for this was failing
because we were using a before callback to ensure that.

Using a validation instead fixes this and seems to be the right way
given that appears to be a validation.

Another sollution would be to use the before_validation callback. But
this seems to be mainly meant preperational operations.

  http://api.rubyonrails.org/classes/ActiveRecord/Callbacks.html#module-ActiveRecord::Callbacks-label-before_validation-2A+returning+statements