Replace `!ref(...)` with `ref(...) eq ''`
From the Perl documentation for the `ref` function:

"If the operand is not a reference, then the empty string will be
returned. An empty string will only be returned in this situation. `ref`
is often useful to just test whether a value is a reference, which can
be done by comparing the result to the empty string. It is a common
mistake to use the result of `ref` directly as a truth value: this goes
wrong because `0` (which is false) can be returned for a reference."