Check if the database exists by its own call in Rails 7.2
It used to rely on the call Rake::Task['db:version'].invoke that
was raising an exception if the database did not exist.

This was changed in Rails 7.2, and it does not raise exceptions anymore.
There is a proper task to check if the database exists though.
Make use of it since Rails 7.2